+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/adp-tooling
 
+## 0.12.12
+
+### Patch Changes
+
+-   Updated dependencies [108336f]
+    -   @sap-ux/inquirer-common@0.4.1
+
 ## 0.12.11
 
 ### Patch Changes